fix: hover trigger on selection layer
