+++ src/dispatch/mouse.ts
@@ -4,7 +4,6 @@ import { changeFocus } from './focus';
 import { changeSelectionRange } from './selection';
 import {
   Store,
-  Side,
   Dimension,
   Selection,
   Viewport,
@@ -15,6 +14,13 @@ import {
 import { getRowRangeWithRowSpan } from '../helper/rowSpan';
 import { getChildColumnRange } from '../query/selection';
 import { findIndexByRowKey } from '../query/data';
+import {
+  findColumnIndexByPosition,
+  findRowIndexByPosition,
+  getColumnNameRange,
+  getPositionFromBodyArea,
+  ElementInfo
+} from '../query/mouse';
 
 export function setNavigating({ focus }: Store, navigating: boolean) {
   focus.navigating = navigating;
@@ -22,7 +28,7 @@ export function setNavigating({ focus }: Store, navigating: boolean) {
 
 type OverflowType = -1 | 0 | 1;
 
-interface ScrollData {
+export interface ScrollData {
   scrollLeft: number;
   scrollTop: number;
 }
@@ -31,14 +37,6 @@ interface OverflowInfo {
   x: OverflowType;
   y: OverflowType;
 }
-
-type ElementInfo = {
-  side: Side;
-  top: number;
-  left: number;
-} & ScrollData;
-
-type ViewInfo = DragData & ScrollData;
 
 interface BodySize {
   bodyWidth: number;
@@ -53,52 +51,6 @@ interface ContainerPosition {
 type EventInfo = DragData & {
   shiftKey: boolean;
 };
-
-function getPositionFromBodyArea(pageX: number, pageY: number, dimension: Dimension) {
-  const {
-    offsetLeft,
-    offsetTop,
-    tableBorderWidth,
-    cellBorderWidth,
-    headerHeight,
-    summaryHeight,
-    summaryPosition
-  } = dimension;
-  const adjustedSummaryHeight = summaryPosition === 'top' ? summaryHeight : 0;
-
-  return {
-    x: pageX - offsetLeft,
-    y:
-      pageY -
-      (offsetTop + headerHeight + adjustedSummaryHeight + cellBorderWidth + tableBorderWidth)
-  };
-}
-
-function getTotalColumnOffsets(widths: { [key in Side]: number[] }, cellBorderWidth: number) {
-  const totalWidths = [...widths.L, ...widths.R];
-  const offsets = [0];
-  for (let i = 1, len = totalWidths.length; i < len; i += 1) {
-    offsets.push(offsets[i - 1] + totalWidths[i - 1] + cellBorderWidth);
-  }
-
-  return offsets;
-}
-
-function getScrolledPosition(
-  { pageX, pageY, scrollLeft, scrollTop }: ViewInfo,
-  dimension: Dimension,
-  leftSideWidth: number
-) {
-  const { x: bodyPositionX, y: bodyPositionY } = getPositionFromBodyArea(pageX, pageY, dimension);
-  const scrollX = bodyPositionX > leftSideWidth ? scrollLeft : 0;
-  const scrolledPositionX = bodyPositionX + scrollX;
-  const scrolledPositionY = bodyPositionY + scrollTop;
-
-  return {
-    x: scrolledPositionX,
-    y: scrolledPositionY
-  };
-}
 
 function judgeOverflow(
   { x: containerX, y: containerY }: ContainerPosition,
@@ -214,49 +166,6 @@ function getFocusCellPos(store: Store) {
     pageX: left + focusCellWidth + (side === 'L' ? 0 : offsetLeft),
     pageY: offsets[rowIndex] + heights[rowIndex] + headerHeight
   };
-}
-
-function findColumnIndexByPosition(store: Store, viewInfo: ViewInfo) {
-  const { dimension, columnCoords } = store;
-  const { widths, areaWidth } = columnCoords;
-  const totalColumnOffsets = getTotalColumnOffsets(widths, dimension.cellBorderWidth);
-  const scrolledPosition = getScrolledPosition(viewInfo, dimension, areaWidth.L);
-
-  return findOffsetIndex(totalColumnOffsets, scrolledPosition.x);
-}
-
-function findRowIndexByPosition(store: Store, viewInfo: ViewInfo) {
-  const { dimension, columnCoords, rowCoords } = store;
-  const { areaWidth } = columnCoords;
-  const scrolledPosition = getScrolledPosition(viewInfo, dimension, areaWidth.L);
-
-  return findOffsetIndex(rowCoords.offsets, scrolledPosition.y);
-}
-
-function getColumnNameRange(
-  store: Store,
-  dragStartData: DragData,
-  dragData: DragData,
-  elementInfo: ElementInfo
-) {
-  const {
-    column: { allColumns }
-  } = store;
-  const { scrollTop, scrollLeft } = elementInfo;
-
-  const { pageX: startPageX, pageY: startPageY } = dragStartData;
-  const { pageX: endPageX, pageY: endPageY } = dragData;
-
-  const startViewInfo = { pageX: startPageX, pageY: startPageY, scrollTop, scrollLeft };
-  const endViewInfo = { pageX: endPageX, pageY: endPageY, scrollTop, scrollLeft };
-
-  const startColumnIndex = findColumnIndexByPosition(store, startViewInfo);
-  const endColumnIndex = findColumnIndexByPosition(store, endViewInfo);
-
-  const { name: startColumnName } = allColumns[startColumnIndex];
-  const { name: endColumnName } = allColumns[endColumnIndex];
-
-  return [startColumnName, endColumnName];
 }
 
 export function selectionEnd({ selection }: Store) {
